WIP for an AWS access log parser
During the last bit of emitting the metrics, it came up that statsd
is a live line protocol only and doesn't allow for backfill which
this was relying on for telegraf to handle aggreagations.

Parking this here in case anyone wants to resume work on it later
and will use a hapi plugin for auth-server instead.